==> 01_python_bronze.ipynb <==
import sys
import os

current_dir = os.path.dirname(os.path.abspath("/dags/scripts"))
sys.path.append(os.path.abspath(os.path.join(current_dir, '..', '..')))

import pandas as pd
import os
from datetime import datetime

os.makedirs('data/bronze', exist_ok=True)
os.makedirs('data/silver', exist_ok=True)
os.makedirs('data/gold', exist_ok=True)
print("Estrutura de pastas verificada.")

# 2. Carregar dados da fonte original
try:
    df_raw = pd.read_csv('data/source/twitter_sentiment.csv', 
                         header=None, 
                         encoding='latin-1') 
    
    # Renomear as colunas
    df_raw.columns = ['Sentimento_Bruto', 'ID', 'Data_Bruta', 'FLAG', 'USER', 'Texto_Bruto']
    
    print(f"Dados carregados: {df_raw.shape[0]} linhas, {df_raw.shape[1]} colunas")

except Exception as e:
    print(f"ERRO DE LEITURA: {e}")

# 3. Adicionar informações de metadados
df_raw['data_ingestao'] = datetime.now()
df_raw['fonte_arquivo'] = 'twitter_sentiment.csv'

print("\nPrimeiras linhas (com metadados):")
print(df_raw.head())

# 4. Salvar na camada Bronze
df_raw.columns = ['Sentimento_Bruto', 'Data_Bruta', 'Texto_Bruto', 'FLAG', 'USER', 'TEXTO_ORIGINAL', 'data_ingestao', 'fonte_arquivo']

output_path = 'data/bronze/dados_brutos.csv' 
df_raw.to_csv(output_path, index=False, header=True)

print(f"\nDados brutos salvos com sucesso na camada Bronze: {output_path}")
